requirements.txt
Django==1.7
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-export-xls==0.1.1
django-toolbelt==0.0.1
gunicorn==19.1.1
newrelic==2.28.0.26
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
xlwt==0.7.5
